requirements.txt
aiogram==2.23.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.9.0
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
imageio==2.33.0
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.31.0
requests==2.28.1
rich==13.7.0
rpds-py==0.13.1
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchutils==0.0.4
torchvision==0.16.1+cu118
tornado==6.3.3
traitlets==5.13.0
triton==2.1.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.13
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.10
yarl==1.9.3
zipp==3.17.0
